requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
gTTS==2.5.1
idna==3.6
multidict==6.0.5
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.0
yarl==1.9.4
